--- packages/ui/src/components/shogi-match/dnd/index.ts
@@ -1,17 +1,0 @@
-/**
- * 将棋盤 編集モード DnD モジュール
- *
- * ref + rAF 方式でパフォーマンス最適化された DnD 実装
- */
-
-// コンポーネント
-// コンテキスト
-export { DragGhost } from "./DragGhost";
-// ドロップロジック
-export { applyDropResult } from "./dropLogic";
-// ヒットテスト
-// 型定義
-export type { DropResult } from "./types";
-export type { PieceDndController as DndController } from "./usePieceDnd";
-// Hooks
-export { usePieceDnd } from "./usePieceDnd";
