+++ packages/ui/src/components/shogi-match/layouts/MobileLayout.tsx
@@ -28,7 +28,9 @@ import { MobileSettingsActions } from "../components/MobileSettingsActions";
 import { MobileSettingsSheet } from "../components/MobileSettingsSheet";
 import { MoveDetailBottomSheet } from "../components/MoveDetailBottomSheet";
 import { PassButton } from "../components/PassButton";
-import { useMatchSettings, useMatchState, useNavigation } from "../contexts";
+import { useMatchSettings } from "../contexts/MatchSettingsContext";
+import { useMatchState } from "../contexts/MatchStateContext";
+import { useNavigation } from "../contexts/NavigationContext";
 import type { DisplaySettings } from "../types";
 import type { KifMove as FullKifMove } from "../utils/kifFormat";
 import type { KifMoveData } from "../utils/kifParser";
